Finished calculating maskArr. Started Level.cpp. Need to calculate shiftArr

+++ pageTable.hpp
@@ -14,7 +14,7 @@ class PageTable
         Level* rootLevel;
         int levelCount;
         long unsigned int *bitMask;        // make sure to use new
-        long unsigned int *bitShift;       // make sure to use new
+        int *bitShift;       // make sure to use new
         int *entryCount;
         unsigned int virtualAddressToPageNum(unsigned int, unsigned int, unsigned int);
